api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
numpy==2.0.1
opencv-python-headless==4.10.0.84
Pillow==10.4.0
